Adding optimization to eliminate many "Position" animations.
If position is animated but anchor isn't, we can convert it to an Offset animation - which is what we do now. That saves us having to create an ExpressionAnimation to relate the anchor and position to the offset.
Start of CompositionSubGraph abstraction in the translator to allow the decision about whether a sub-graph is a shape or a visual to be put off until there is more context available. This will allow us to use Opacity for visiblity animation in some circumstances, which will enable some more TreeReducer optimizations to kick in.