Add more detail into the code generation comment at the top of files.
The goal is to be able to identify which version of a JSON file, and what switches and build were used. This should help reduce confusion when there is a lot of iterating going on with a design.
There are some changes to the .sln file that were a side effect of updating to a new VS version - not sure what they do.